Upgraded PyTorch version to 0.4.1 and switched to Python 3.7.

+++ canny.py
@@ -2,9 +2,6 @@ from scipy.misc import imread, imsave
 import torch
 from torch.autograd import Variable
 from net_canny import Net
-from matplotlib import pyplot as plt
-import matplotlib
-import numpy as np
 
 
 def canny(raw_img, use_cuda=False):
@@ -24,21 +21,12 @@ def canny(raw_img, use_cuda=False):
 
     imsave('gradient_magnitude.png',grad_mag.data.cpu().numpy()[0,0])
     imsave('thin_edges.png', thresholded.data.cpu().numpy()[0, 0])
-    imsave('final.png', thresholded.data.cpu().numpy()[0, 0]  > 0.0)
+    imsave('final.png', (thresholded.data.cpu().numpy()[0, 0] > 0.0).astype(float))
     imsave('thresholded.png', early_threshold.data.cpu().numpy()[0, 0])
 
-    # plt.imshow(thresholded.data.cpu().numpy()[0, 0] > 0.0, interpolation='None', cmap='gray')
-    # plt.show()
 
 if __name__ == '__main__':
     img = imread('fb_profile.jpg') / 255.0
-
-    # img = np.array([    [0, 0, 0, 0, 0],
-    #                     [0, 0, 0, 0, 0],
-    #                     [0, 0, 0, 0, 0],
-    #                     [0, 0, 1, 0, 0],
-    #                     [0, 0, 0, 0, 0]     ])
-    # img = np.stack([img,img,img]).transpose([1,2,0])
 
     # canny(img, use_cuda=False)
     canny(img, use_cuda=True)
